Refactor to remove top-level exports
