+++ examples/extract-people-names.ts
@@ -1,6 +1,7 @@
 import 'dotenv/config';
 
-import { ChatModel, createAIExtractFunction } from '@dexaai/dexter';
+import { ChatModel } from '@dexaai/dexter/model';
+import { createAIExtractFunction } from '@dexaai/dexter/prompt';
 import { z } from 'zod';
 
 /** A function to extract people names from text. */
