+++ package.json
@@ -40,8 +40,6 @@
       "import": "./dist/utils/index.js"
     }
   },
-  "main": "./dist/index.js",
-  "types": "./dist/index.d.ts",
   "sideEffects": false,
   "files": [
     "dist"
