+++ src/index.ts
@@ -1,6 +1,3 @@
-// this file is only used as the entry point for documentation generation
+// This is intentionally empty, but here to make bundlers happy.
 
-export * from './model/index.js';
-export * from './prompt/index.js';
-export * from './telemetry/index.js';
-export * from './utils/index.js';
+export {};
